backend/requirements.txt
flask
tensorflow
pillow
numpy
flask-cors